rtos/os-c-wrapper.c: osDelay() with uint64_t cast
- without cast the 32-bits always cut the value

+++ src/rtos/os-c-wrapper.cpp
@@ -2766,7 +2766,7 @@ osDelay (uint32_t millisec)
     }
 
   result_t res = sysclock.sleep_for (
-      clock_systick::ticks_cast (millisec * 1000u));
+      clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)));
 
   if (res == ETIMEDOUT)
     {
@@ -2812,7 +2812,7 @@ osWait (uint32_t millisec)
     }
 
   result_t res = sysclock.wait_for (
-      clock_systick::ticks_cast (millisec * 1000u));
+      clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)));
 
   // TODO: return events
   if (res == ETIMEDOUT)
@@ -2883,7 +2883,7 @@ osTimerStart (osTimerId timer_id, uint32_t millisec)
     }
 
   result_t res = (reinterpret_cast<rtos::timer&> (*timer_id)).start (
-      clock_systick::ticks_cast (millisec * 1000u));
+      clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)));
 
   if (res == result::ok)
     {
@@ -3059,7 +3059,7 @@ osSignalWait (int32_t signals, uint32_t millisec)
   else
     {
       res = this_thread::timed_sig_wait (
-          (flags::mask_t) signals, clock_systick::ticks_cast (millisec * 1000u),
+          (flags::mask_t) signals, clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)),
           (flags::mask_t*) &event.value.signals);
     }
 
@@ -3168,7 +3168,7 @@ osMutexWait (osMutexId mutex_id, uint32_t millisec)
   else
     {
       ret = (reinterpret_cast<rtos::mutex&> (*mutex_id)).timed_lock (
-          clock_systick::ticks_cast (millisec * 1000u));
+          clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)));
       // osErrorTimeoutResource:
     }
 
@@ -3359,7 +3359,7 @@ osSemaphoreWait (osSemaphoreId semaphore_id, uint32_t millisec)
   else
     {
       res = (reinterpret_cast<rtos::semaphore&> (*semaphore_id)).timed_wait (
-          clock_systick::ticks_cast (millisec * 1000u));
+          clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)));
       if (res == ETIMEDOUT)
         {
           return 0;
@@ -3643,7 +3643,7 @@ osMessagePut (osMessageQId queue_id, uint32_t info, uint32_t millisec)
         }
       res = (reinterpret_cast<message_queue&> (*queue_id)).timed_send (
           (const char*) &info, sizeof(uint32_t),
-          clock_systick::ticks_cast (millisec * 1000u), 0);
+          clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)), 0);
       // osOK, osErrorTimeoutResource, osErrorParameter
     }
 
@@ -3736,7 +3736,7 @@ osMessageGet (osMessageQId queue_id, uint32_t millisec)
         }
       res = (reinterpret_cast<message_queue&> (*queue_id)).timed_receive (
           (char*) &event.value.v, sizeof(uint32_t),
-          clock_systick::ticks_cast (millisec * 1000u), NULL);
+          clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)), NULL);
       // result::event_message when message;
       // result::event_timeout when timeout;
     }
@@ -3869,7 +3869,7 @@ osMailAlloc (osMailQId mail_id, uint32_t millisec)
           return nullptr;
         }
       ret = (reinterpret_cast<memory_pool&> (mail_id->pool)).timed_alloc (
-          clock_systick::ticks_cast (millisec * 1000u));
+          clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)));
     }
 #pragma GCC diagnostic pop
   return ret;
@@ -4019,7 +4019,7 @@ osMailGet (osMailQId mail_id, uint32_t millisec)
         }
       res = (reinterpret_cast<message_queue&> (mail_id->queue)).timed_receive (
           (char*) &event.value.p, sizeof(void*),
-          clock_systick::ticks_cast (millisec * 1000u), NULL);
+          clock_systick::ticks_cast ((uint64_t)(millisec * 1000u)), NULL);
       // osEventMail for ok, osEventTimeout
     }
